pyproject.toml
[tool.poetry]
name = "api-django"
version = "0.1.0"
description = ""
authors = ["lucassantos-dev <jl.lucassantos098@gmail.com>"]
readme = "README.md"
packages = [{include = "api_django"}]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.14.0"
djoser = "^2.2.0"
django-cors-headers = "^4.1.0"
python-dotenv = "^1.0.0"
django = "^4.2.3"
pillow = "^10.0.0"
django-extensions = "^3.2.3"
werkzeug = "^2.3.6"
django-werkzeug-debugger-runserver = "^0.3.1"
django-sslserver = "^0.22"
gunicorn = "^20.1.0"
pandas = "^2.0.3"
drf-yasg = "^1.21.7"
requests = "^2.31.0"
holidays = "^0.29"
tqdm = "^4.65.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
